requirements.txt
ango==4.2.7
Pillow==10.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
